# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A complete REST API for managing multiple WhatsApp sessions using Go, Fiber, PostgreSQL, and whatsmeow library.\n\n## Authentication\nAll API endpoints (except /health and /swagger/*) require API key authentication.\nProvide your API key in the `Authorization` header.",
        "title": "zpwoot - WhatsApp Multi-Session API",
        "contact": {
            "name": "zpwoot Support",
            "url": "https://github.com/your-org/zpwoot",
            "email": "support@zpwoot.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/chatwoot/config": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new Chatwoot integration configuration. This enables synchronization between Wameow and Chatwoot. Requires API key authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatwoot"
                ],
                "summary": "Create Chatwoot configuration",
                "parameters": [
                    {
                        "description": "Chatwoot configuration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateChatwootConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Chatwoot configuration created successfully",
                        "schema": {
                            "$ref": "#/definitions/CreateChatwootConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or parameters",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateChatwootConfigRequest": {
            "type": "object",
            "required": [
                "accountId",
                "apiKey",
                "url"
            ],
            "properties": {
                "accountId": {
                    "type": "string",
                    "example": "1"
                },
                "apiKey": {
                    "type": "string",
                    "example": "chatwoot-api-key-123"
                },
                "inboxId": {
                    "type": "string",
                    "example": "1"
                },
                "url": {
                    "type": "string",
                    "example": "https://chatwoot.example.com"
                }
            }
        },
        "CreateChatwootConfigResponse": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string",
                    "example": "1"
                },
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "chatwoot-config-123"
                },
                "inboxId": {
                    "type": "string",
                    "example": "1"
                },
                "url": {
                    "type": "string",
                    "example": "https://chatwoot.example.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API key directly (no Bearer prefix required). Example: dev-api-key-12345",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}